requirements.txt
python-telegram-bot[job-queue]==21.9
python-dotenv==1.0.0
requests==2.31.0
instagrapi==2.1.3
Pillow==10.2.0
moviepy==1.0.3
tiktok-uploader==0.0.6